Extract Team component into a separate file.
This extraction will make it possible to test the Team component in
isolation.
